Move script files corresponding to CrabNet publication
